Address review: rename DiscoveryError.UpstreamError and move DTOs to AuthModels
- Rename GitHubError → UpstreamError (accurate for both proxy 5xx and upstream failures)
- Remove redundant TaskCanceledException from exception filters (it inherits OperationCanceledException)
- Move ProxyConfigResponse, DiscoveredTenant, DiscoveryError, DiscoveryResult from AuthProxyClient.cs to AuthModels.cs
